refactor: rename users module to useradmin
The `user` (entity + repository + TenantUserDetails) and `users`
(Tenant-Owner administration) packages collide one character apart
and force prose every time to disambiguate. Rename `users` to
`useradmin` to make the admin-of-users role legible at the import
line and in the package tree.

Modulith verifier passes against the renamed module. The DB table
`users` and the URL prefix `/manage/t/{slug}/users/...` are unchanged.
