+++ docs/reference/architecture.md
@@ -308,7 +308,7 @@ There are two tenant-scoped login surfaces — an OAuth2 end-user login at `/t/{
 
 Order matters: the password-change check fires **before** OAuth2-resume so a User with an expired password cannot complete an authorize flow before updating it. New policies are added by registering an `@Bean PostLoginIntent` with an `@Order` value; user-supplied intents are prepended to the defaults via `ObjectProvider#orderedStream()`.
 
-**Forced password change.** Two trigger paths share one orchestrator. On a fresh login, the `passwordChangeRequired()` intent (above) catches a User whose `must_change_password` flag is set and redirects them to the surface's change-password URL. Inside the management console — for example, after an admin clicks **Reset password** mid-session — `PasswordChangeRequiredInterceptor` (in `users`) catches subsequent requests and does the same redirect. A single package-private `PasswordChangeController` in `auth.login` binds both URL prefixes (`/manage/t/{slug}/change-password` and `/t/{slug}/change-password`) and dispatches per request via `TenantUrlScheme.slugFrom(uri)`; it delegates to `TenantPasswordChangeFlow` (also in `auth.login`) for validation, persistence, and OAuth2-authorize-resume. The resume target is always tenant-prefixed under `/t/{slug}/oauth2/authorize` regardless of which surface the User changed their password on, because the authorize endpoint only lives on the OAuth2 surface.
+**Forced password change.** Two trigger paths share one orchestrator. On a fresh login, the `passwordChangeRequired()` intent (above) catches a User whose `must_change_password` flag is set and redirects them to the surface's change-password URL. Inside the management console — for example, after an admin clicks **Reset password** mid-session — `PasswordChangeRequiredInterceptor` (in `useradmin`) catches subsequent requests and does the same redirect. A single package-private `PasswordChangeController` in `auth.login` binds both URL prefixes (`/manage/t/{slug}/change-password` and `/t/{slug}/change-password`) and dispatches per request via `TenantUrlScheme.slugFrom(uri)`; it delegates to `TenantPasswordChangeFlow` (also in `auth.login`) for validation, persistence, and OAuth2-authorize-resume. The resume target is always tenant-prefixed under `/t/{slug}/oauth2/authorize` regardless of which surface the User changed their password on, because the authorize endpoint only lives on the OAuth2 surface.
 
 **Account lockout.** A pre-authentication check in `TenantAuthProvider` rejects logins for Users whose `locked_until` is in the future, throwing `LockedException` *before* the password is verified — so a locked account does not leak whether the supplied password was right. `LoginAttemptTracker` is an `@EventListener` on `AuthenticationFailureEvent` / `AuthenticationSuccessEvent` that increments `failed_login_attempts` on failure, sets `locked_until` once the threshold (default 5) is reached, and clears both columns on success. Tenant Owners reset the counters via `POST /manage/t/{slug}/users/{userId}/unlock` on the User detail page (`UserManagementController.unlock()` → `UserAdministrationService.unlockAccount(...)` → `AccountUnlockedEvent` for the audit log). Threshold and lockout window are configured via `LockoutProperties`.
 
@@ -592,7 +592,7 @@ Each direct sub-package of `com.stucray.limen` is one application module. As of 
 | `system` | Cross-tenant System Admin controllers (tenant suspend/unsuspend/delete, system-admin tenant-create) |
 | `tenant` | `Tenant` entity + repository, `TenantStatus`, `TenantScope` (the per-request `ScopedValue`) |
 | `user` | `User` entity + `UserRepository` + `TenantUserDetails` (the Spring Security `UserDetails` adapter wrapping `User` + `Tenant`) |
-| `users` | Tenant-Owner administration of Users: `UserAdministrationService`, `PasswordChangeRequiredInterceptor` (the change-password form itself is hosted by `auth.login.PasswordChangeController`) |
+| `useradmin` | Tenant-Owner administration of Users: `UserAdministrationService`, `PasswordChangeRequiredInterceptor` (the change-password form itself is hosted by `auth.login.PasswordChangeController`) |
 | `web` | Top-level web routes: `RootController` (landing) and `RedirectLoginController` (slug-aware `/login` forwarder) |
 
 **Three module-shaping decisions worth calling out:**
