+++ src/main/java/com/stucray/limen/management/web/ManagementWebConfig.java
@@ -1,6 +1,6 @@
 package com.stucray.limen.management.web;
 
-import com.stucray.limen.users.PasswordChangeRequiredInterceptor;
+import com.stucray.limen.useradmin.PasswordChangeRequiredInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
