+++ src/main/java/com/stucray/limen/useradmin/PasswordChangeRequiredInterceptor.java
@@ -1,4 +1,4 @@
-package com.stucray.limen.users;
+package com.stucray.limen.useradmin;
 
 import com.stucray.limen.user.TenantUserDetails;
 import jakarta.servlet.http.HttpServletRequest;
