+++ src/main/java/com/stucray/limen/useradmin/UserAdministrationService.java
@@ -1,4 +1,4 @@
-package com.stucray.limen.users;
+package com.stucray.limen.useradmin;
 
 import com.stucray.limen.audit.events.AccountUnlockedEvent;
 import com.stucray.limen.audit.events.PasswordChangedEvent;
