+++ src/test/java/com/stucray/limen/user/EmailUniquenessIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.stucray.limen.user;
 
 import com.stucray.limen.TestcontainersConfiguration;
-import com.stucray.limen.users.UserAdministrationService;
+import com.stucray.limen.useradmin.UserAdministrationService;
 import com.stucray.limen.tenant.Tenant;
 import com.stucray.limen.provisioning.TenantProvisioningService;
 import com.stucray.limen.tenant.TenantRepository;
