+++ src/test/java/com/stucray/limen/useradmin/UserDetailControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.stucray.limen.users;
+package com.stucray.limen.useradmin;
 
 import com.stucray.limen.clients.CreateClientCommand;
